fix: invoice, place order, add notification

+++ src/main/java/com/startup/tasteflowbe/dto/request/CreateWarehouseRequestDTO.java
@@ -6,8 +6,8 @@ import java.util.List;
 // DTO chính để tạo phiếu yêu cầu
 @Data
 public class CreateWarehouseRequestDTO {
-    private Integer warehouseId;
-    private Integer createdBy;
+    private Long warehouseId;
+    private Long createdBy;
     private String notes;
     private List<RequestItemDTO> items;
 }
